refactor(evidence): split finding merge out of the CI cycle
Move merge, identity, and row validation into a leaf both findings and
ci.evidence import. Split line helpers without a flag interpreter; keep
one Unreleased Fixed bullet per issue.
