+++ src/mergecraft/analyzers/supply_chain.py
@@ -204,7 +204,7 @@ def _osv_package_names_by_file(raw: str) -> dict[str, str]:
     package_names: dict[str, str] = {}
     try:
         payload = json.loads(raw)
-    except json.JSONDecodeError:
+    except ValueError:
         return package_names
     for result in payload.get("results") or []:
         if not isinstance(result, dict):
